+++ CHANGELOG.md
@@ -6,6 +6,16 @@
   `webhookVerifier` support for rotated credentials and webhook-forwarding
   infrastructure such as Vercel Connect.
 
+## v3.24.1
+
+### `@liveblocks/react`
+
+- Add new hook `useMutableStorage()` to get direct access to the mutable Storage
+  root. See
+  [docs](https://liveblocks.io/docs/api-reference/liveblocks-react#useMutableStorage).
+- Fix thread resolved/unresolved status being reverted under some conditions.
+  (Thanks @VihaanAgarwal for the contribution!)
+
 ## v3.24.0
 
 This release introduces `LiveText` (beta), a collaborative rich-text data